refactor: refactor adventure message imports and reorganize component extensions
