+++ surf-api-core/surf-api-core-api/src/main/kotlin/dev/slne/surf/surfapi/core/api/messages/adventure/bossbar-extension.kt
@@ -1,6 +1,5 @@
 package dev.slne.surf.surfapi.core.api.messages.adventure
 
-import dev.slne.surf.surfapi.core.api.messages.buildText
 import dev.slne.surf.surfapi.core.api.messages.builder.SurfComponentBuilder
 import dev.slne.surf.surfapi.core.api.util.mutableObjectSetOf
 import net.kyori.adventure.bossbar.BossBar
