+++ surf-api-core/surf-api-core-api/src/main/kotlin/dev/slne/surf/surfapi/core/api/messages/adventure/resource-pack-extensions.kt
@@ -1,6 +1,5 @@
 package dev.slne.surf.surfapi.core.api.messages.adventure
 
-import dev.slne.surf.surfapi.core.api.messages.buildText
 import dev.slne.surf.surfapi.core.api.messages.builder.SurfComponentBuilder
 import dev.slne.surf.surfapi.core.api.util.mutableObjectListOf
 import kotlinx.coroutines.Dispatchers
